add defaultFS in sqlTaskContext
